+++ mocket/__init__.py
@@ -2,7 +2,7 @@ from mocket.async_mocket import async_mocketize
 from mocket.entry import MocketEntry
 from mocket.mocket import Mocket
 from mocket.mocketizer import Mocketizer, mocketize
-from mocket.ssl import FakeSSLContext
+from mocket.ssl.context import FakeSSLContext
 
 __all__ = (
     "async_mocketize",
